+++ src/dij.rs
@@ -1,9 +1,13 @@
 #![allow(clippy::type_complexity, dead_code)]
 use super::*;
-use graph::{DirectedGraph, DirectedGraphTrait, Graph2, UndirectedAdjGraph};
+use crate::utils::min_max;
+use graph::{DirectedGraph, DirectedGraphTrait, Graph2};
 use haversine::haversine_m_fpair;
 use itertools::Itertools;
 use ordered_float::OrderedFloat;
+use radix_heap::RadixHeapMap;
+use sorted_slice_store::SortedSliceMap;
+use std::cmp::Reverse;
 use std::collections::BinaryHeap;
 
 #[derive(Clone, Debug, Default, clap::ValueEnum, PartialEq)]
@@ -14,73 +18,39 @@ pub enum SplitPathsMethod {
 }
 
 /// Does a single Dijkstra search from start_idx to all vertexes.
-/// return a hashmap for each other vertex in the graph, with the previous node to go to (if
-/// applicable) and the total distance along the path
-fn dij_single(
+/// The results are in the prev_dist variable, hashmap for each other vertex in the graph, with the
+/// previous node to go to (if applicable) and the total distance along the path By passing in as a
+/// &mut argument, the allocated memory can be reused between runs
+pub fn dij_single(
     start_idx: i64,
-    edges: &UndirectedAdjGraph<i64, f32>,
-) -> HashMap<i64, (Option<i64>, OrderedFloat<f32>)> {
-    trace!(
-        "dij_single started. start_idx {} edges.len() {}",
-        start_idx,
-        edges.len()
-    );
-    let mut prev_dist = HashMap::new();
-    prev_dist
-        .try_reserve(edges.num_vertexes())
-        .expect("Unable to reserve enough space for Dij. alg.");
-    prev_dist.insert(start_idx, (None, OrderedFloat(0.)));
+    edges: &Graph2,
+    edge_lengths: &SortedSliceMap<(i64, i64), u64>,
+    prev_dist: &mut HashMap<i64, (Option<i64>, u64)>,
+) {
+    prev_dist.clear();
+    prev_dist.reserve(edges.num_vertexes());
+    prev_dist.extend(edges.vertexes().map(|nid| (*nid, (None, u64::MAX))));
 
-    let mut frontier = BinaryHeap::new();
-    // hack to just store negative distance so the shortest distance is the largest number
-    frontier.push((OrderedFloat(-0.), start_idx));
+    prev_dist.insert(start_idx, (None, 0));
 
-    let mut this_dist;
-    while let Some((mut curr_dist, curr_id)) = frontier.pop() {
-        curr_dist *= -1.;
-        //trace!(
-        //    "Current frontier. id {:>13} curr dist {:>8}, currently shortest known: {:>10} frontier.len() {:>3}",
-        //    curr_id,
-        //    curr_dist,
-        //    prev_dist[&curr_id].1,
-        //    frontier.len()
-        //);
+    let mut frontier = RadixHeapMap::new();
+    frontier.push(Reverse(0), start_idx);
+
+    while let Some((Reverse(curr_dist), curr_id)) = frontier.pop() {
         if curr_dist > prev_dist[&curr_id].1 {
-            // already found a shorter
             continue;
         }
-        for neighbor in edges
-            .neighbors(&curr_id)
-            .map(|(i, d)| (i, OrderedFloat(*d)))
-        {
-            this_dist = neighbor.1 + curr_dist;
-            prev_dist
-                .entry(*neighbor.0)
-                .and_modify(|(prev_id, dist)| {
-                    if this_dist < *dist {
-                        *prev_id = Some(curr_id);
-                        *dist = this_dist;
-                        frontier.push((-this_dist, *neighbor.0));
-                    }
-                })
-                .or_insert_with(|| {
-                    frontier.push((-this_dist, *neighbor.0));
-                    (Some(curr_id), this_dist)
-                });
-            //trace!("Now prev_dist.len {}", prev_dist.len());
+        for neighbour_idx in edges.neighbours(&curr_id) {
+            let this_dist =
+                curr_dist + edge_lengths.get(&min_max(curr_id, *neighbour_idx)).unwrap();
+            let old_value = prev_dist.get_mut(neighbour_idx).unwrap();
+            if this_dist < old_value.1 {
+                old_value.0 = Some(curr_id);
+                old_value.1 = this_dist;
+                frontier.push(Reverse(this_dist), *neighbour_idx);
+            }
         }
     }
-
-    let results = prev_dist;
-
-    trace!(
-        "dij_single finished. start_idx {} edges.len() {} results.len() {}",
-        start_idx,
-        edges.len(),
-        results.len()
-    );
-
-    results
 }
 
 /// Does a single A* search from start_idx to a set of other vertexes
@@ -89,7 +59,9 @@ pub(crate) fn paths_one_to_many<'a>(
     targets: &'a [(i64, (OrderedFloat<f64>, OrderedFloat<f64>))],
     nodeid_pos: &'a impl NodeIdPosition,
     edges: &'a Graph2,
+    edge_lengths: impl Into<Option<&'a SortedSliceMap<(i64, i64), f64>>>,
 ) -> impl ParallelIterator<Item = ((i64, i64), Vec<i64>)> + 'a {
+    let edge_lengths: Option<&'a SortedSliceMap<(i64, i64), f64>> = edge_lengths.into();
     let (start_idx, _start_pos) = start;
     assert!(edges.contains_vertex(start_idx));
     assert!(
@@ -105,7 +77,7 @@ pub(crate) fn paths_one_to_many<'a>(
 
     targets
         .par_iter()
-        .map(move |target| path_one_to_one(start, *target, nodeid_pos, edges))
+        .map(move |target| path_one_to_one(start, *target, nodeid_pos, edges, edge_lengths))
 }
 
 /// Does a single A* search from start_idx to another
@@ -114,7 +86,9 @@ pub(crate) fn path_one_to_one<'a>(
     target: (i64, (OrderedFloat<f64>, OrderedFloat<f64>)),
     nodeid_pos: &'a impl NodeIdPosition,
     edges: &'a Graph2,
+    edge_lengths: impl Into<Option<&'a SortedSliceMap<(i64, i64), f64>>>,
 ) -> ((i64, i64), Vec<i64>) {
+    let edge_lengths: Option<&'a SortedSliceMap<(i64, i64), f64>> = edge_lengths.into();
     let (start_idx, start_pos) = start;
     let (target_idx, target_pos) = target;
     let mut frontier = BinaryHeap::new();
@@ -153,10 +127,14 @@ pub(crate) fn path_one_to_one<'a>(
             continue;
         }
         for neighbor_idx in edges.neighbours(&curr_id) {
-            let edge_len = haversine_m_fpair(
-                (*curr_pos.0, *curr_pos.1),
-                nodeid_pos.get(neighbor_idx).unwrap(),
-            );
+            let edge_len = if let Some(edge_lengths) = edge_lengths {
+                *edge_lengths.get(&min_max(curr_id, *neighbor_idx)).unwrap()
+            } else {
+                haversine_m_fpair(
+                    (*curr_pos.0, *curr_pos.1),
+                    nodeid_pos.get(neighbor_idx).unwrap(),
+                )
+            };
             this_path_dist = curr_path_dist + OrderedFloat(edge_len);
             //info!("Looking at neighbour id {}, which is {} from start", neighbor_idx, this_path_dist);
             best_dist_prev