+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Edge Betweenness Centrality can be calculated (`--betweenness-centrality`)
+  * Includes a JOSM map style to show this data
 * New Tag filter `∃key∉value1,value2,value3,…`
 * New option `--min-length N%longest` and `--min-length inclN%total`